+++ next.config.mjs
@@ -1,6 +1,13 @@
 import withMarkdoc from '@markdoc/next.js'
-
+import nextPWA from 'next-pwa'
+import runtimeCaching from 'next-pwa/cache.js'
 import withSearch from './src/markdoc/search.mjs'
+
+const withPWA = nextPWA({
+  dest: 'public',
+  runtimeCaching,
+  disable: process.env.NODE_ENV === 'development',
+})
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -10,9 +17,9 @@ const nextConfig = {
   output: 'export',
   distDir: 'dist',
   images: {
-    unoptimized: true
+    unoptimized: true,
   },
   pageExtensions: ['js', 'jsx', 'md'],
 }
 
-export default withSearch(withMarkdoc({ schemaPath: './src/markdoc' })(nextConfig))
+export default withSearch(withMarkdoc({ schemaPath: './src/markdoc' })(withPWA(nextConfig)))